p2p-node/requirements.txt
websockets==12.0
PyYAML==6.0.1